backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
python-jose[cryptography]
bcrypt
python-dotenv
python-multipart
python-dateutil
slowapi
xhtml2pdf
google-auth==2.29.0
google-auth-oauthlib==1.2.0
httpx==0.27.0
email-validator
razorpay
